requirements.txt
aiohttp==3.6.2
py-cord==1.7.3
dnspython==1.16.0
emoji==0.5.4
isodate==0.6.0
motor==2.1.0
natural==0.2.0
parsedatetime==2.6
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.14.0
websockets==9.1
lavalink
DiscordUtils
psutil